EntityConfigurations Finished

+++ Entities/Concrete/TableModels/Order.cs
@@ -6,10 +6,10 @@ namespace Entities.Concrete.TableModels
     {
         public int CartId { get; set; }
         public int ShippingAdressId { get; set; }
-        public int SubTotal { get; set; }
-        public short Delivery { get; set; }
-        public short Commision { get; set; }
-        public int GrandTotal { get; set; }
+        public decimal SubTotal { get; set; }
+        public float Delivery { get; set; }
+        public float Commision { get; set; }
+        public float GrandTotal { get; set; }
         public bool IsDelivery { get; set; }
         public virtual Cart Cart { get; set; }
         public virtual ShippingAdress ShippingAdress { get; set; }
